Remove stray help info from mediaelement.js
We noticed the following string on or below some of the pictures
in the LoveInAction "exhibition":
"Use Left/Right Arrow keys to"
Turns out this is a help/info message generated by mediaelement.js,
but it was nowhere near the video or audio files. The display
CSS property was set to none to hide it.